+++ src/imp/interrupt/armv6m.rs
@@ -12,6 +12,14 @@ Generated asm:
 #[cfg(not(portable_atomic_no_asm))]
 use core::arch::asm;
 
+#[cfg_attr(
+    portable_atomic_no_cfg_target_has_atomic,
+    cfg(any(test, portable_atomic_no_atomic_cas))
+)]
+#[cfg_attr(
+    not(portable_atomic_no_cfg_target_has_atomic),
+    cfg(any(test, not(target_has_atomic = "ptr")))
+)]
 pub(super) use core::sync::atomic;
 
 pub(super) type State = u32;
@@ -23,11 +31,11 @@ pub(super) fn disable() -> State {
     // SAFETY: reading the priority mask register and disabling interrupts are safe.
     // (see module-level comments of interrupt/mod.rs on the safety of using privileged instructions)
     unsafe {
-        // Do not use `nomem` and `readonly` because prevent subsequent memory accesses from being reordered before interrupts are disabled.
         asm!(
             "mrs {primask}, PRIMASK", // primask = PRIMASK
             "cpsid i",                // PRIMASK.PM = 1
             primask = out(reg) primask,
+            // Do not use `nomem` and `readonly` because prevent subsequent memory accesses from being reordered before interrupts are disabled.
             options(nostack, preserves_flags),
         );
     }
@@ -45,11 +53,10 @@ pub(super) unsafe fn restore(prev_primask: State) {
     // and we've checked that interrupts were enabled before disabling interrupts.
     unsafe {
         // This clobbers the entire PRIMASK register. See msp430.rs to safety on this.
-        //
-        // Do not use `nomem` and `readonly` because prevent preceding memory accesses from being reordered after interrupts are enabled.
         asm!(
             "msr PRIMASK, {prev_primask}", // PRIMASK = prev_primask
             prev_primask = in(reg) prev_primask,
+            // Do not use `nomem` and `readonly` because prevent preceding memory accesses from being reordered after interrupts are enabled.
             options(nostack, preserves_flags),
         );
     }